pyproject.toml
[project]
name = "zeitgeist-engine"
version = "2.0.0"
description = "AI-powered global news digest that captures the world's attention every 4 hours"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Zeitgeist Team" }]
keywords = ["news", "ai", "llm", "digest", "zeitgeist", "bluesky", "mastodon"]

dependencies = [
    # Data Collection
    "google-cloud-bigquery>=3.14.0",
    "atproto>=0.0.50",
    "Mastodon.py>=1.8.1",
    "pytrends>=4.9.2",
    "trafilatura>=1.6.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",

    # Processing
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.4",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "scipy>=1.11.0",

    # Vector Database
    "qdrant-client>=1.7.0",

    # LLM APIs (Multi-Model Support)
    "google-generativeai>=0.3.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",

    # Configuration
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "requests>=2.31.0",
    "aiohttp>=3.9.0",

    # Logging
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
zeitgeist = "src.main:run_pipeline"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
